refactor: rename class and export
