+++ viewer/api-entry-points/core.ts
@@ -11,6 +11,7 @@ export * from '../packages/api/';
 export {
   CAMERA_MANAGER_EVENT_TYPE_LIST,
   CameraControlsOptions,
+  DebouncedCameraStopEventTrigger,
   DefaultCameraManager,
   CameraManagerEventType,
   CameraManagerHelper,
