+++ viewer/packages/camera-manager/index.ts
@@ -7,5 +7,6 @@ export { StationaryCameraManager } from './src/StationaryCameraManager';
 export { CameraManagerHelper } from './src/CameraManagerHelper';
 export { CameraManager } from './src/CameraManager';
 export { ComboControls, ComboControlsOptions } from './src/ComboControls';
+export { DebouncedCameraStopEventTrigger } from './src/utils/DebouncedCameraStopEventTrigger';
 
 export * from './src/types';
