+++ viewer/packages/camera-manager/src/utils/DebouncedCameraStopEventTrigger.ts
